[trello.com/c/qdDT0Ic1] Markdown optimization
